requirements.txt
flask
line-bot-sdk==2.4.1
firebase-admin
python-dotenv
